作業項目と無関係のロジックを別クラスに分離。
